Assignment6-CRUD_FastAPI/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
typing
httpx
pytest